Renamed network_clustering to network_util because why not. KNN caching, clustering, layout, and plotting. Now with more-labeled dendrograms!
Minor change to common to remove some commented out stuff. I don't think
we want to compute coordinates on loading at this point. Compute offline
or on demand.

tissue_analysis only has two functions in it at the moment.

The Tissue Exploration Notebook eXtravaganza (TENX) is a bit of a work in progress but it should allow for rapid reclustering of a given dataset.